src/main/java/com/eaton/hug4bugs/testgoodpractice/badcodebase/model/Person.java
package com.eaton.hug4bugs.testgoodpractice.badcodebase.model;

import java.util.Date;
import java.util.Set;

public class Person {

    private String firstname;

    private String lastname;

    private Date birthDay;

    private Set<Vehicle> vehicles;

    public Person(String firstname, String lastname) {
        this.firstname = firstname;
        this.lastname = lastname;
    }

    public String getFirstname() {
        return firstname;
    }

    public void setFirstname(String firstname) {
        this.firstname = firstname;
    }

    public String getLastname() {
        return lastname;
    }

    public void setLastname(String lastname) {
        this.lastname = lastname;
    }

    public Date getBirthDay() {
        return birthDay;
    }

    public void setBirthDay(Date birthDay) {
        this.birthDay = birthDay;
    }

    public void setVehicles(Set<Vehicle> vehicles) {
        this.vehicles = vehicles;
    }

    public Set<Vehicle> getVehicles() {
        return vehicles;
    }
}
